chore: move cmd
