+++ package.json
@@ -16,6 +16,7 @@
   "devDependencies": {
     "@oclif/dev-cli": "^1",
     "@oclif/plugin-command-snapshot": "^3.1.3",
+    "@oclif/test": "1.2.9",
     "@salesforce/cli-plugins-testkit": "^2.3.0",
     "@salesforce/dev-config": "^3.0.1",
     "@salesforce/dev-scripts": "^2.0.2",
